move test and fix add files.
